Enable the weather example to function with non-English languages (#3013)
This is a quick suggestion by a non-native English speaker! When passing
non-English prompt like "今日の東京の天気は？" (this means "What's the weather in
Tokyo today?" in Japanese language), the example app fails with the
following error:
```
ERROR [2025-03-18 11:35:00.589 +0900] (weatherTool): Failed tool execution
    runId: "weatherAgent"
    description: "Get current weather for a location"
    args: {
      "location": "東京"
    }
    error: {}
```

This pull request resolves this issue not only for Japanese but also
other non-English languages.