Fix symlinks and externals in bundler (#3071)
Rollup wasn't handling pnpm symlinks properly because of the
`preserveSymlinks` argument. We use this to make sure we get the actual
package name instead of weird symlink folder in our babel transforms.
The optimizer doesn't do any transforms so we do not need it.

As an addition we're now handling externals correctly.