Kenny/mastra 2742 migrate nextra from v3 to v4 (#3310)
Docs Migration to Nextra 4.

Nextra 4 includes support for:
1. NextJS App Router
2. Turbopack
3. Tailwind V4
4. A new search primitive, PageFind

Nextra new version deprecates some custom components and so we cannot
modify the `table-of-contents` component like we did in `v3` The only
settings provided are for modifying the toc itself rather than replacing
it:

https://nextra.site/docs/docs-theme/built-ins/layout#table-of-contents-toc

With this, the search should be better for custom components.