feat: bundle design tokens as CJS in playground UI for tailwind usage (#3577)
## Description

The goal is to be able to use the design tokens in the playground-ui AND
in the cli/playground one. The problem is that playground-ui is bundled
as ESM and the tailwind config of cli/playground only supports cjs.

This PR introduces an additional "token" entry to the playground-ui vite
config file so that we can compile it to ESM and CJS.

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes
- [x] I have generated a changeset for this PR