feat(@mastra/pg): schema support (#3463)
## Description

This is a non-breaking change that allows `@mastra/pg`'s storage and
vector module to be used with a custom schema.

Motivation for me personally is that I want to give Mastra its own
schema and have it manage it and not interfere with my main schema,
which I manage with a migration system.

Here's how you can pass the schema to `PostgresStore`:

```ts
export const mastraStorage = new PostgresStore({
  connectionString: process.env.DATABASE_URL,
  schema: "myschema", // new
});

```

And `PgVector`:

```ts
export const mastraVector = new PgVector(process.env.DATABASE_URL) // this still works
export const mastraVector = new PgVector({
  connectionString: process.env.DATABASE_URL,
  schema: "myschema",
}); // can also pass an object with your custom schema
```

If this is a feature the maintainers are OK with merging, I can put in
some more work and add tests, changeset and update docs.

Thank you!

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] I have generated a changeset for this PR

---------

Co-authored-by: NikAiyer <nick.aiyer@gmail.com>