feat: send mcp name on error (#3664)
## Description

<img width="844" alt="Screenshot 2025-04-16 at 09 58 16"
src="https://github.com/user-attachments/assets/2fc0051d-09f3-44ec-b783-1904e46cf137"
/>


If there a connection error on any MCP tool, we log the name of which
tool had the issue. But the error that we then throw doesn't include
that metadata. That makes it harder for the consumer to know which of
those tools is the one that failed.

This PR includes the name of the tool that failed in the error message.

## Type of Change
- [X] Code refactoring

## Checklist

- [X] I have performed a self-review of my own code
- [x] I have generated a changeset for this PR