feat(core): tool schema compatibility layer for model providers (#4027)
## Description

This PR aim's to add a tool compatibility layer to fix the
inconsistencies between models from different providers. Models may not
be able to support all json schema properties (such as some openai
reasoning models), as well as other models support the property but seem
to ignore it.

When writing your own tools, you have control to tweak the schema as you
choose to make it fit with the model you're using, however when using
tools someone else has written (ie from MCP), you don't have that
control. The author of the MCP tools may have written the tools using a
different model that interprets the schema differently.

The intention here is to make all tools behave similarly regardless of
the model used and have this as invisible to the Mastra user as
possible. To do this we remove the incompatible schema pieces and append
the instructions to the end of the provided tool description for the
agent to use.

To work around this we introduce a `ToolCompatibility` class interface. 
This class has
- a `shouldApply` method which will dictate the logic around when this
compatibility class should be applied to the tool.
- a `getSchemaTarget` function which will be used to determine the
schema format.
- a `processZodType` which handles the transformations of the schema
- and finally, multiple helper methods to help with the processing and
transformations.

6 compatibility classes are shipped with this
- google gemini
- anthropic
- openai
- openai reasoning models
- deepseek
- meta

<!-- Provide a brief description of the changes in this PR -->

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [x] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] I have generated a changeset for this PR

---------

Co-authored-by: Tyler Barnes <tylerdbarnes@gmail.com>
Co-authored-by: Daniel Lew <daniellew@Daniels-MacBook-Pro.local>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>