[MASTRA-3439] Working Memory tool call fix (#4303)
…t content

## Description

<!-- Provide a brief description of the changes in this PR -->
This PR fixes an issue where if the agent sent a message before updating
working memory, that message would not be saved in the db.

We were skipping messages completely that had workingmemory in the
message content, but instead we are now filtering the message content to
remove references to workingmemory.

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->
#4221 

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works