[MASTRA-3091] added runtime context to vector-query and graph-rag tools (#4433)
## Description

<!-- Provide a brief description of the changes in this PR -->
This PR adds an optional field to all for the use of runtime context for
the vector-query tool and graph-rag tool. When used, users can inject
the indexName, vectorStoreName, topK, filter and other parameters into
the tool when using the agent. This allows users to specify their own
values for these fields without needing to create different tools or
modify their prompts.
## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works