[docs] remove mention of evaluating tools in evals (#4446)
## Description

I don't see an example of how to get reference to tool calls in a Metric
or MastraAgentJudge or EvaluationResult.
And in
[this](https://discord.com/channels/1309558646228779139/1362166435799105626/1362166435799105626)
discord message it seems verified this is not supported.
So I think it's best to remove the mention of "Assess how effectively an
agent uses external tools or APIs" in the Eval docs (for the time-being
at least)

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [x] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works