feat(@mastra/core, @mastra/memory): Store UIMessages in memory (#4436)
Before this PR we were storing CoreMessages in memory. The problem is
the conversion from UIMessage->CoreMessage is lossy, so when messages
are queried from memory and returned to the UI, essential information
got lost.

This PR uses the MessageList class as a translation layer so we can take
in CoreMessage, UIMessage, and Mastra db messages stored originally as
CoreMessage (MastraMessageV1) or as UIMessage (MastraMessageV2). The
list can then output in any of those four formats Core/UI/V1/V2.

This allows us to start storing UIMessages without breaking anyone who's
already using memory and has previously stored CoreMessages. When those
are remembered they will be approximated to UIMessages. Any new input
messages will be stored as UIMessages if input as such, or otherwise
converted as close as possible.

This allows us to support AI SDK v5 next!