quotes table names for vector and storage to allow for camelcase (#5148)
## Description

<!-- Provide a brief description of the changes in this PR -->
This PR fixes an issue with PG where camelCase names (as well as any
indexName that contained uppercase letters) weren't being saved
properly. The indexNames were not being quoted, so the case of the index
names were not being preserved. This PR changes this to quote the table
and schema names to allow for different cases.

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works