feat(core:agent): Pass thread metadata to agent stream/generate (#5108)
## Description

Adds ability to pass thread metadata to `agent.stream()` and
`agent.generate()`. This will update the current thread metadata or
create a new thread with the metadata if the thread does not exist yet.

This PR also reworks the arguments for those two functions to simplify
them.

Before:
```
{
  ...otherArgs
  threadId
  resourceId
  memoryConfig
}
```

After
```
{
   ...otherArgs
   memory: {
      thread: { id, metadata, ... } 
      resource: 123 // we'll need this to be an object in the future too
      options: MemoryConfig
  }
}
```

This ensures that it's easier to follow what arguments are related to
memory and declutters the top level type.

The old args are still there as we need backwards compatibility, but
I've added a deprecation notice on it so we can remove them in the
future.
<!-- Provide a brief description of the changes in this PR -->

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: Daniel Lew <daniellew@Daniels-MacBook-Pro.local>