fix(schema-compat): allow for object.passthrough (#5442)
## Description

Initially we didn't allow for passthrough as it's not a good pattern for
writing tools as LLM's do much better with explicitly typed schemas.
However some people try to use MCP servers that use this and because of
that it fails every time. This PR allows for passthroughs to be used
(except with OpenAI reasoning models as those throw errors when
passthroughs are used). using an MCP that uses object.passthrough will
not give you good performance, but at least they will work sometimes.

<!-- Provide a brief description of the changes in this PR -->

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [x] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: Daniel Lew <daniellew@Daniels-MacBook-Pro.local>