[MASTRA-4118] AgentNetwork Message Display (#5474)
## Description

<!-- Provide a brief description of the changes in this PR -->
This PR fixes two issues with AgentNetwork UI when using loopStream:
- the results of agent calls are not being displayed in the UI on
refresh. this is due to the message list concatenating assistant
messages together, and we are not displaying all of the parts of a
system message. This causes some messages to disappear on refresh.
- subsequent agent calls aren't being saved to the messages table. This
is because when new assistant messages are added to the message list,
they are appended to the latest assistant message, causing them not to
be saved when draining messages.

TODO:
We now save all internal agent network messages to mastra messages, but
now we need to figure out how to display the messages we want. Currently
all messages generated during AgentNetwork loop stream are being
displayed, when only the results of the calls should show up.

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: Taofeeq Oluder <oluderutaofeeq@gmail.com>
Co-authored-by: Tony Kovanen <rase-@users.noreply.github.com>
Co-authored-by: Tony Kovanen <tony.kovanen@gmail.com>