fix(mcp): handle jsonSchema $defs and $ref property (#5761)
## Description

MCP Tools from non-mastra tools such as fast-mcp, favour using the $defs
and $ref property in the json schema. Zod transformations can't handle
this. However there is a popular package
https://www.npmjs.com/package/json-schema-ref-parser that flattens the
json schema from the $defs and $ref property so that it's able to be
transformed into zod.

This was tested with a python fast mcp server as we can't test it with
Mastra tools because we don't allow passing in a JSONSchema
<!-- Provide a brief description of the changes in this PR -->

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works